Building Agentic AI Systems/requirements.txt
openai>=1.30.0
numpy>=1.26.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
requests>=2.31.0